Cleanup document.py: Moved all views and viewlets in to a seperate file and right directory.
