+++ opengever/document/browser/view_templates/view.pt
